// types/report/leaves/leave-types.ts
export interface LeaveReports{
    id: number;
    employee: string;
    department: string;
    work_status: string;
    leave_type: string;
    leave_status: string;
    leave_start_date: string;
    leave_end_date: string;
    leave_duration: string;
    leave_reason: string;
    leave_allocated_days: string;
    leave_used_days: string;
    leave_carry_forward_days: string;
    leave_remaining_days: string;
}